Torch, Skeleton Enemy, Heal Spell

+++ Dark Souls Like/Assets/Scripts/CharacterStats.cs
@@ -11,8 +11,8 @@ namespace NC
         public int currentHealth;
 
         public int staminaLevel = 10;
-        public int maxStamina;
-        public int currentStamina;
+        public float maxStamina;
+        public float currentStamina;
 
         public bool isDead;
     }
